+++ inflearnJpa1/src/main/java/jpabook/jpashop/repository/OrderRepository.java
@@ -106,12 +106,4 @@ public class OrderRepository {
                         " join fetch o.delivery d", Order.class
         ).getResultList();
     }
-
-    public List<OrderSimpleQueryDto> findOrderDtos() {
-        return em.createQuery(
-                        "select new jpabook.jpashop.repository.OrderSimpleQueryDto(o.id, m.name, o.orderDate, o.status, d.address) from Order o" +
-                                " join o.member m" +
-                                " join o.delivery d", OrderSimpleQueryDto.class)
-                .getResultList();
-    }
 }
